checking connection with frontend

+++ backend/controllers/product.controller.js
@@ -57,12 +57,18 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        console.error("Invalid Product ID");
+        return res.status(404).json({ success: false, message: "Invalid Product Id" });
+    }
+    
    try{
     await Product.findByIdAndDelete(id);
     res.status(200).json({success: true, message: "Product deleted!"});
 
    }catch(error){
     console.log("error in deleting product:", error.message);
-    res.status(404).json({success: false, message: "Product not found!"});
+    res.status(500).json({success: false, message: "Sever Error!"});
    }
 };
